Refactor Governor SDK into Modular Structure
